support for transparent/sizescale setting on modelattach
Former-commit-id: aeb8c484c4bcbc91ece81dfce968489eb6757c2b
Former-commit-id: 04e16c2128c6201229a7cd8a2e0365fd6e804877

+++ src/game/hud.cpp
@@ -638,7 +638,7 @@ namespace hud
     }
     ICOMMAND(0, getpointer, "ii", (int *i, int *j), result(getpointer(*i, *j)));
 
-    void drawindicator(int weap, int x, int y, int s, bool secondary)
+    void drawindicator(int weap, int x, int y, float s, bool secondary)
     {
         int millis = lastmillis-game::focus->weaplast[weap];
         float r = 1, g = 1, b = 1, amt = 0;
